.ts > .js

+++ server/src/modules/book/resolver/addRemove.ts
@@ -1,5 +1,5 @@
-import collection from '@models/collection.ts'
-import updateCollection from '@services/redis/updateCollection.ts'
+import collection from '@models/collection.js'
+import updateCollection from '@services/redis/updateCollection.js'
 import type { User } from '@type/models/user.d.ts'
 
 const addRemove = async (_: null, args: { author_key: string[], cover_edition_key: string, cover_i: number, title: string, author_name: string[] }, context: { user: User }) => {
@@ -12,10 +12,7 @@ const addRemove = async (_: null, args: { author_key: string[], cover_edition_ke
             cover_edition_key,
             cover_i
         }).lean()
-        if (bookCollection) {
-            await collection.findByIdAndDelete(bookCollection._id)
-            await updateCollection('collection', user)
-        }
+        if (bookCollection) await collection.findByIdAndDelete(bookCollection._id)
         else {
             await collection.create({
                 user_id: user._id,
@@ -26,8 +23,8 @@ const addRemove = async (_: null, args: { author_key: string[], cover_edition_ke
                 author_name,
                 created: new Date()
             })
-            await updateCollection('collection', user)
         }
+        await updateCollection('collection', user)
         return true
     } catch (e) {
         throw e
